Movement and player char made lol
I made a movement script. it used an efficient event based system. it controls dash, movement and the direction the player is looking/ aiming. I made some "layers" to control the order in which different things were layered/ presented by the camera. player is on layer 1, summons and enemies are on layer 2 and 3 and the map is layer 4 so that it is always beneath us. I added a gizmo to show us player rotation. it can be turned off by disabling gizmos. I made a sprite for the player. he is in a white scientist jacket with a black top-hat and a special green backback filled with healing spray juice since he is a scientist and support healer. // there is something strange the first few times you dash, afterwards i think its quite smooth. the variables that control dash speed, cooldown and duration are exposed to edit. good luck kneeler i hope you read this while drinking a nice coffee and petting a cat.